Change: Move add_sys_path, unload_module and ensure_unload_module to helpers
These functions are useful not only for testing purposes.
